+++ src/main/java/com/github/sgoeschl/freemarker/cli/model/Documents.java
@@ -52,6 +52,10 @@ public class Documents implements Closeable {
         return documents.isEmpty();
     }
 
+    public Document getFirst() {
+        return documents.get(0);
+    }
+
     public List<Document> getList() {
         return new ArrayList<>(documents);
     }