+++ src/test/java/com/github/sgoeschl/freemarker/cli/ExamplesTest.java
@@ -26,7 +26,7 @@ import static org.junit.Assert.assertTrue;
 
 public class ExamplesTest extends AbstractMainTest {
 
-    private static final int MIN_OUTPUT_SIZE = 128;
+    private static final int MIN_OUTPUT_SIZE = 1;
 
     @Test
     public void shouldRunInfo() throws IOException {
@@ -103,6 +103,13 @@ public class ExamplesTest extends AbstractMainTest {
     }
 
     @Test
+    public void shouldRunInteractiveTemplateExamples() throws IOException {
+        assertValid(execute("-i ${JsonPathTool.parse(Documents.first).read(\"$.info.title\")} site/sample/json/swagger-spec.json"));
+        assertValid(execute("-i ${XmlTool.parse(Documents.first)[\"recipients/person[1]/name\"]} site/sample/xml/recipients.xml"));
+        assertValid(execute("-i ${JsoupTool.parse(Documents.first).select(\"a\")[0]} site/sample/html/dependencies.html"));
+    }
+
+    @Test
     @Ignore("Manual test to check memory consumption and resource handling")
     public void shouldCloseAllResources() throws IOException {
         for (int i = 0; i < 5000; i++) {
@@ -116,6 +123,7 @@ public class ExamplesTest extends AbstractMainTest {
             shouldRunYamlExamples();
             shouldRunXmlExamples();
             shouldRunGrokExamples();
+            shouldRunInteractiveTemplateExamples();
             shouldRunWithExposedEnvironmentVariableExamples();
         }
     }