+++ src/test/java/com/github/sgoeschl/freemarker/cli/ManualTest.java
@@ -24,10 +24,12 @@ public class ManualTest {
     private static final String SPACE = " ";
     // private static final String CMD = "-b ./src/test -t templates/csv/html/transactions.ftl site/sample/csv/transactions.csv";
     // private static final String CMD = "-b ./src/test -l de_AT -Dfoo=bar -t templates/info.ftl site/sample/csv/transactions.csv";
-    private static final String CMD = "-b ./src/test -DFOO=foo -DBAR=bar -l de -t templates/demo.ftl site/sample/csv/transactions.csv";
+    // private static final String CMD = "-b ./src/test -DFOO=foo -DBAR=bar -l de -t templates/demo.ftl site/sample/csv/transactions.csv";
     // private static final String CMD = "-b ./src/test -DFOO=foo -DBAR=bar -t templates/demo.ftl site/sample/csv/transactions.csv";
     // private static final String CMD = "-b ./src/test -Dcsv.out.format=TDF -t templates/csv/transform.ftl site/sample/csv/contract.csv";
     // private static final String CMD = "-t templates/excel/csv/transform.ftl -l de_AT site/sample/excel/test.xlsx";
+    // private static final String CMD = "-i ${JsonPathTool.parse(Documents.get(0)).read(\"$.info.title\")} site/sample/json/swagger-spec.json";
+    private static final String CMD = "-i ${XmlTool.parse(Documents.first)[\"recipients/person[1]/name\"]} site/sample/xml/recipients.xml";
 
     public static void main(String[] args) {
         Main.execute(CMD.split(SPACE));